+++ src/ImageResize/AbstractResize.php
@@ -33,11 +33,6 @@ abstract class AbstractResize
     protected $processed = false;
 
     /**
-     * @var boolean Has the create() method been called?
-     */
-    protected $resized = false;
-
-    /**
      * @var string Filename for new image file
      */
     protected $file = null;
@@ -72,19 +67,18 @@ abstract class AbstractResize
         array $canvas_color = array('r' => 255, 'g' => 255, 'b' => 255)
     ) {
         if ($width < 1) {
-            throw new \InvalidArgumentException('Width not valid, must be greater than 0');
+            throw new \InvalidArgumentException(Helper::ERROR_WIDTH_INVALID);
         }
 
         if ($height < 1) {
-            throw new \InvalidArgumentException('Height not valid, must be greater than 0');
+            throw new \InvalidArgumentException(Helper::ERROR_HEIGHT_INVALID);
         }
 
-        if ($this->colorIndexValid('r', $canvas_color) === false ||
-            $this->colorIndexValid('g', $canvas_color) === false ||
-            $this->colorIndexValid('b', $canvas_color) === false
+        if (Helper::colorIndexValid('r', $canvas_color) === false ||
+            Helper::colorIndexValid('g', $canvas_color) === false ||
+            Helper::colorIndexValid('b', $canvas_color) === false
         ) {
-            throw new \InvalidArgumentException('Canvas colour array invalid, it should contain three indexes, r, g 
-                and b and each should have a value between 0 and 255');
+            throw new \InvalidArgumentException(Helper::ERROR_CANVAS_COLOR_ARRAY_INVALID);
         }
 
         $this->canvas['width'] = $width;
@@ -95,22 +89,61 @@ abstract class AbstractResize
     }
 
     /**
-     * Check to see if the supplied color index is valid
+     * Set a new width value, useful if you are using the resize class within a loop/iterator
      *
-     * @param string $index
-     * @param array $color The color array to check
+     * @param int $width
      *
-     * @return boolean
+     * @return AbstractResize
+     * @throws \Exception
      */
-    private function colorIndexValid(string $index, array $color) : bool
+    public function setWidth(int $width) : AbstractResize
     {
-        if (array_key_exists($index, $color) === true &&
-            $color[$index] >= 0 && $color[$index] <= 255
-        ) {
-            return true;
-        } else {
-            return false;
+        if ($width < 1) {
+            throw new \InvalidArgumentException(Helper::ERROR_WIDTH_INVALID);
         }
+
+        $this->canvas['width'] = $width;
+
+        return $this;
+    }
+
+    /**
+     * Set a new height value, useful if you are using the resize class within a loop/iterator
+     *
+     * @param int $height
+     *
+     * @return AbstractResize
+     * @throws \Exception
+     */
+    public function setHeight(int $height) : AbstractResize
+    {
+        if ($height < 1) {
+            throw new \InvalidArgumentException(Helper::ERROR_HEIGHT_INVALID);
+        }
+
+        $this->canvas['height'] = $height;
+
+        return $this;
+    }
+
+    /**
+     * Set a new canvas color
+     *
+     * @param array $canvas_color Indexed array, r, g, b for color
+     *
+     * @return AbstractResize
+     * @throws \Exception
+     */
+    public function setCanvasColor(array $canvas_color) : AbstractResize
+    {
+        if (Helper::colorIndexValid('r', $canvas_color) === false ||
+            Helper::colorIndexValid('g', $canvas_color) === false ||
+            Helper::colorIndexValid('b', $canvas_color) === false
+        ) {
+            throw new \InvalidArgumentException(Helper::ERROR_CANVAS_COLOR_ARRAY_INVALID);
+        }
+
+        $this->canvas['color'] = $canvas_color;
     }
 
     /**
@@ -127,10 +160,9 @@ abstract class AbstractResize
         if (file_exists($path . $file) === true) {
             $this->source['path'] = $path;
             $this->source['file'] = $file;
-            $this->sourceDimensions();
+            $this->sourceProperties();
         } else {
-            throw new \Exception("File couldn't be found, supplied 
-			destination: '" . $path . $file . "'");
+            throw new \Exception("File couldn't be found in supplied destination: '" . $path . $file . "'");
         }
 
         // Reset the processed status when a new image is loaded
@@ -149,17 +181,20 @@ abstract class AbstractResize
      * @return void
      * @throws \Exception
      */
-    protected function sourceDimensions()
+    protected function sourceProperties()
     {
-        $dimensions = getimagesize($this->source['path'] . $this->source['file']);
+        $properties = Helper::imageProperties($this->source['file'], $this->source['path']);
 
-        $this->source['width'] = intval($dimensions[0]);
-        $this->source['height'] = intval($dimensions[1]);
-        $this->source['aspect_ratio'] = floatval($this->source['width'] / $this->source['height']);
+        if ($properties['width'] !== null) {
+            $this->source['width'] = $properties['width'];
+            $this->source['height'] = $properties['height'];
+            $this->source['aspect_ratio'] = $properties['aspect_ratio'];
+        } else {
+            throw new \Exception(Helper::ERROR_CALL_GETIMAGESIZE);
+        }
 
         if ($this->canvas['width'] > $this->source['width'] || $this->canvas['height'] > $this->source['height']) {
-            throw new \Exception("The options are set to upscale the image, this class 
-             does not support that.");
+            throw new \Exception(Helper::ERROR_UPSCALE);
         }
     }
 
@@ -353,7 +388,7 @@ abstract class AbstractResize
                 ]
             ];
         } else {
-            throw new \Exception('Unable to getInfo(), process() not called');
+            throw new \Exception(Helper::ERROR_CALL_GETINFO);
         }
     }
 
@@ -366,17 +401,17 @@ abstract class AbstractResize
     {
         $this->canvas['canvas'] = imagecreatetruecolor($this->canvas['width'], $this->canvas['height']);
         if ($this->canvas['canvas'] === false) {
-            throw new \Exception('Call to imagecreatetruecolor failed');
+            throw new \Exception(Helper::ERROR_CALL_IMAGECREATETRUECOLOR);
         }
 
         $fill_color = imagecolorallocate($this->canvas['canvas'], $this->canvas['color']['r'],
             $this->canvas['color']['g'], $this->canvas['color']['b']);
         if ($fill_color === false) {
-            throw new \Exception('Call to imagecolorallocate failed');
+            throw new \Exception(Helper::ERROR_CALL_IMAGECOLORALLOCATE);
         }
 
         if (imagefill($this->canvas['canvas'], 0, 0, $fill_color) === false) {
-            throw new \Exception('Call to imagefill failed');
+            throw new \Exception(Helper::ERROR_CALL_IMAGEFILL);
         };
     }
 
@@ -393,7 +428,7 @@ abstract class AbstractResize
             $this->source['height']);
 
         if($result === false) {
-            throw new \Exception('Call to imagecopyresampled failed');
+            throw new \Exception(Helper::ERROR_CALL_IMAGECOPYRESAMPLED);
         }
     }